refactor: 리뷰 내용 반영
- package 명 변경
- 외부에서 사용하지 않는 메소드의 접근제어자 변경
- GeneratedValue 찾는 메소드 내부 변수명 리네임
