+++ src/main/java/persistence/sql/dml/DeleteQueryBuilder.java
@@ -1,8 +1,8 @@
 package persistence.sql.dml;
 
 import persistence.sql.QueryBuilder;
-import persistence.sql.core.Entity;
-import persistence.sql.core.PrimaryKey;
+import persistence.sql.metadata.Entity;
+import persistence.sql.metadata.PrimaryKey;
 
 public class DeleteQueryBuilder extends QueryBuilder {
