+++ src/main/java/persistence/sql/metadata/PrimaryKey.java
@@ -1,4 +1,4 @@
-package persistence.sql.core;
+package persistence.sql.metadata;
 
 public class PrimaryKey {
